requirements.txt
python==3.6.5
pandas==1.1.5
requests==2.25.1
bs4==0.0.1
djangorestframework	==3.12.4